% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_data_baseline.R
\name{add_holiday_effect}
\alias{add_holiday_effect}
\title{add holiday effect ----
add holiday efect to baseline data}
\usage{
add_holiday_effect(data, holiday_data, holiday_effect = 2)
}
\arguments{
\item{splfmt_rds_data}{Simulated or real data}

\item{start_date}{Starting date simulated outbreak.
Date is in the format of 'yyyy-mm-dd'.}

\item{numoutbk}{Ending date of the simulation period.}
}
\value{
A splfmt_rts_data_v1, data.table containing

\describe{
  \item{granularity_time}{Pseudo ID for death events.}
  \item{granularity_geo}{Pseudo ID for death events.}
  \item{wday}{Pseudo ID for death events.}
}
}
\description{
add holiday effect ----
add holiday efect to baseline data
}
\details{

}
