% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farrington_algorithm.R
\name{farrington_algorithm}
\alias{farrington_algorithm}
\title{Farrington algorithm}
\usage{
farrington_algorithm(data, reweights = 1, remove.highcounts = 0, is_day = T)
}
\arguments{
\item{reweights}{Number (greater or equal to 0) of residual reweights adjusting for previous outbreaks (default: 1; 1 reweight)}

\item{remove.highcounts}{Number between 0 and 1 of fraction of high counts to be removed from prediction, to remove impact of earlier outbreaks (default: 0)}

\item{d_train}{Training data.table}

\item{datasetPredict}{Prediction data.table}

\item{sign.level}{Significance level for the prediction intervals (default: 0.05)}

\item{isDaily}{Is it daily data or weekly data?}

\item{v}{Version (Not in use)}

\item{weeklyDenominatorFunction}{sum or mean - should the denominator be summed or meaned over time}

\item{uuid}{uuid}
}
\description{
Description:
}
\details{
Input (arguments):
dataset: data frame with columns for number of cases, covariates and dates per day
datecol: name of date-column (default: 'Dato')
predinterval: length of prediction interval (default: 30; last 30 days)
historical.data.years: number (should be greater or equal to at least 3) of full years of background data (default: 5)
mod.pred.window: number (greater or equal to 0) of days window between datasets used for modelling and prediction(default: 90)
reweights: number (greater or equal to 0) of residual reweights adjusting for previous outbreaks (default: 1; 1 reweight)
remove.pandemic.year: true/false (default: false; keep 2009 data)
remove.highcounts: number between 0 and 1 of fraction of high counts to be removed from prediction, to remove impact of earlier outbreaks (default: 0)
sign.level: significance level for the prediction intervals (default: 5%)
}
