% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farrighton_flexible.R
\name{farrington_algorithm_create_alerts}
\alias{farrington_algorithm_create_alerts}
\title{Farrington Flexible algorithm}
\usage{
farrington_algorithm_create_alerts(
  data,
  method = "farrington_flexible",
  p = 10,
  b = 5,
  freq = 365,
  w = 3,
  weeks_excluded_n = 23,
  offset = F,
  reweights = 1,
  remove.highcounts = 0,
  is_day = T,
  s = 2.58
)
}
\arguments{
\item{data}{Training splfmt_rds data object}

\item{p}{Number of levels in the factor for seasonality adjustment}

\item{b}{number of years to used for baseline}

\item{w}{Window's half-size}

\item{weeks_excluded_n}{Number of past weeks to remove in the calculation}
}
\description{
This function is based on farringtonFlexible {surveillance} according to the procedure developed by Farrington et al. (1996) and  Noufaily et al. (2012).
Seasonality is adjusted for by including in the model a p-level factor as described in Noufaily et al. (2012)
}
