% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farrington_flexible_algorithm.R
\name{farrington_flexible_se_gamma_space}
\alias{farrington_flexible_se_gamma_space}
\title{Calculate Farrington Flexible SE in gamma space}
\usage{
farrington_flexible_se_gamma_space(
  pred,
  phi,
  alpha = NULL,
  z = NULL,
  skewness.transform = "none"
)
}
\arguments{
\item{pred}{Point estimate}

\item{phi}{Dispersion}

\item{alpha}{Not used}

\item{z}{Not used}

\item{skewness.transform}{"none"/"1/2","2/3"}
}
\description{
Calculate Farrington Flexible SE in gamma space
}
