% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_thresholds.R
\name{prediction_thresholds}
\alias{prediction_thresholds}
\title{Prediction thresholds}
\usage{
prediction_thresholds(object, newdata, alpha = 0.05, z = NULL, ...)
}
\arguments{
\item{object}{Object}

\item{newdata}{New data}

\item{alpha}{Two-sided alpha (e.g 0.05)}

\item{z}{Similar to \code{alpha} (e.g. z=1.96 is the same as alpha=0.05)}

\item{...}{dots}
}
\description{
Prediction thresholds
}
