% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasipoisson_alerts.R
\name{quasipoisson_algorithm_create_alerts}
\alias{quasipoisson_algorithm_create_alerts}
\title{Quasipoisson Algorithm Create Alerts}
\usage{
quasipoisson_algorithm_create_alerts(
  data,
  b = 5,
  offset = F,
  reweights = 1,
  remove.highcounts = 0,
  is_day = T,
  s = 2.58
)
}
\arguments{
\item{data}{Training splfmt_rds data object}

\item{reweights}{Number (greater or equal to 0) of residual reweights adjusting for previous outbreaks (default: 1; 1 reweight)}

\item{remove.highcounts}{Number between 0 and 1 of fraction of high counts to be removed from prediction, to remove impact of earlier outbreaks (default: 0)}

\item{is_day}{Significance level for the prediction intervals (default: 0.05)}
}
\description{
Description:
Applys a surveillance algorithm based on a quasi-poisson regression
model to the selected data. The difference from the Farrington algorithm is in how
seasonality is accounted for (here it is adjusted for, in Farrington it is removed
by design.
}
