% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_term_trend.R
\name{short_term_trend}
\alias{short_term_trend}
\title{Determine the short term trend}
\usage{
short_term_trend(
  x,
  numerator,
  denominator = NULL,
  trend_days = 14,
  remove_last_days = 0,
  forecast_days = trend_days
)
}
\arguments{
\item{x}{Data object}

\item{numerator}{Character of name of numerator}

\item{denominator}{Character of name of denominator (optional)}

\item{trend_days}{Number of days you want to check the trend}

\item{remove_last_days}{Number of days you want to remove at the end (due to unreliable data)}

\item{forecast_days}{Number of days you want to forecast into the future}
}
\description{
Determine the short term trend
}
