% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_data_baseline.R
\name{add_holiday_effect}
\alias{add_holiday_effect}
\title{Holiday effect ----}
\usage{
add_holiday_effect(data, holiday_data, holiday_effect = 2)
}
\arguments{
\item{data}{A splfmt_rds data object}

\item{holiday_data}{dates}

\item{holiday_effect}{Ending date of the simulation period.}
}
\value{
A splfmt_rts_data_v1, data.table containing
}
\description{
The effect of public holiday on a time series of daily counts
}
