% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_term_trend.R
\name{short_term_trend}
\alias{short_term_trend}
\title{Determine the short term trend}
\usage{
short_term_trend(
  x,
  numerator,
  denominator = NULL,
  prX = 100,
  trend_dates = 42,
  remove_last_dates = 0,
  forecast_dates = trend_dates,
  trend_isoyearweeks = ceiling(trend_dates/7),
  remove_last_isoyearweeks = ceiling(remove_last_dates/7),
  forecast_isoyearweeks = trend_isoyearweeks,
  numerator_naming_prefix = "from_numerator",
  denominator_naming_prefix = "from_denominator",
  statistics_naming_prefix = "universal",
  remove_training_data = FALSE
)
}
\arguments{
\item{x}{Data object}

\item{numerator}{Character of name of numerator}

\item{denominator}{Character of name of denominator (optional)}

\item{prX}{If using denominator, what scaling factor should be used for numerator/denominator?}

\item{trend_dates}{Number of dates you want to check the trend}

\item{remove_last_dates}{Number of dates you want to remove at the end (due to unreliable data)}

\item{forecast_dates}{Number of dates you want to forecast into the future}

\item{trend_isoyearweeks}{Same as trend_dates, but used if granularity_geo=='isoyearweek'}

\item{remove_last_isoyearweeks}{Same as remove_last_dates, but used if granularity_geo=='isoyearweek'}

\item{forecast_isoyearweeks}{Same as forecast_dates, but used if granularity_geo=='isoyearweek'}

\item{numerator_naming_prefix}{"from_numerator", "generic", or a custom prefix}

\item{denominator_naming_prefix}{"from_denominator", "generic", or a custom prefix}

\item{statistics_naming_prefix}{"universal" (one variable for trend status, one variable for doubling dates), "from_numerator_and_prX" (If denominator is NULL, then one variable corresponding to numerator. If denominator exists, then one variable for each of the prXs)}

\item{remove_training_data}{Boolean. If TRUE, removes the training data (i.e. 1:(trend_dates-1) or 1:(trend_isoyearweeks-1)) from the returned dataset.}
}
\description{
Determine the short term trend
}
