% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_term_trend.R
\name{short_term_trend}
\alias{short_term_trend}
\title{Determine the short term trend}
\usage{
short_term_trend(x, trend_days = 14, remove_last_days = 0)
}
\arguments{
\item{x}{Data object}

\item{trend_days}{Number of days you want to check the trend}

\item{remove_last_days}{Number of days you want to remove at the end (due to unreliable data)}
}
\description{
Determine the short term trend
}
