% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_data_baseline.R
\name{simulate_baseline_data}
\alias{simulate_baseline_data}
\title{Simulate (daily) baseline data}
\usage{
simulate_baseline_data(
  start_date,
  end_date,
  seasonal_pattern_n,
  weekly_pattern_n,
  param_list = list(alpha, beta, gamma1, gamma3, gamma4, phi, shift)
)
}
\arguments{
\item{start_date}{Starting date of the simulation period.
Date is in the format of 'yyyy-mm-dd'.}

\item{end_date}{Ending date of the simulation period.}

\item{param_list}{List of parameters to control the baseline simulation.}

\item{seasonal_pattern}{}

\item{weekly_pattern}{}
}
\value{
A splfmt_rts_data_v1, data.table containing

\describe{
  \item{granularity_time}{Pseudo ID for death events.}
  \item{granularity_geo}{Pseudo ID for death events.}
  \item{wday}{Pseudo ID for death events.}
}
}
\description{
Simulates a time series of counts based on the Poisson/Negative Binomial model as described in Noufaily et al. (2019)
}
\details{

}
\examples{
start_date <- '2018-01-01'
end_date <- '2019-12-31'

 baseline  <- simulate_baseline_data(
start_date = as.Date("2012-01-01")
end_date = as.Date("2019-12-31")
seasonal_pattern_n = 1
weekly_pattern_n = 1
param_list = list(alpha = 3, beta = 0,gamma1 = 0.8, gamma2 = 0.6, gamma3 = 0.8, gamma4 = 0.4, phi = 4,shift = 29 )
)
}
