% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_data_baseline.R
\name{simulate_seasonal_outbreak_data}
\alias{simulate_seasonal_outbreak_data}
\title{Simulate (daily) seasonal outbreak data}
\usage{
simulate_seasonal_outbreak_data(
  data,
  week_season_start = 40,
  week_season_peak = 4,
  week_season_end = 20,
  num_season_outbreak = 1,
  baseline_param_list = list(alpha, beta, gamma1, gamma3, gamma4, phi, shift),
  outbreak_param_list = list(m)
)
}
\arguments{
\item{splfmt_rds_data}{Simulated or real data}

\item{start_date}{Starting date simulated outbreak.
Date is in the format of 'yyyy-mm-dd'.}

\item{numoutbk}{Ending date of the simulation period.}
}
\value{
A splfmt_rts_data_v1, data.table containing

\describe{
  \item{granularity_time}{Pseudo ID for death events.}
  \item{granularity_geo}{Pseudo ID for death events.}
  \item{wday}{Pseudo ID for death events.}
}
}
\description{
Simulate (daily) seasonal outbreak data
}
\details{

}
\examples{
start_date <- '2018-01-01'
end_date <- '2019-12-31'

 baseline  <- simulate_baseline_data(
 start_date = as.Date("2012-01-01")
 end_date = as.Date("2019-12-31")
 seasonal_pattern_n = 1
 weekly_pattern_n = 2
baseline_param_list = list(alpha = 2, beta = 0.0001,gamma1 = 0.2, gamma2 = 0.5, gamma3 = 0.5, gamma4 = 0.2, phi = 4, shift = 6 )

outbreak_param_list = list(m=10)

)
}
