% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_data_baseline.R
\name{simulate_spike_outbreak_data}
\alias{simulate_spike_outbreak_data}
\title{simulate spike outbreak data ----
Simulate (daily) spike outbreak data}
\usage{
simulate_spike_outbreak_data(data, n_sp_outbreak = 1, m)
}
\arguments{
\item{splfmt_rds_data}{Simulated or real data}

\item{start_date}{Starting date simulated outbreak.
Date is in the format of 'yyyy-mm-dd'.}

\item{numoutbk}{Ending date of the simulation period.}
}
\value{
A splfmt_rts_data_v1, data.table containing

\describe{
  \item{granularity_time}{Pseudo ID for death events.}
  \item{granularity_geo}{Pseudo ID for death events.}
  \item{wday}{Pseudo ID for death events.}
}
}
\description{
simulate spike outbreak data ----
Simulate (daily) spike outbreak data
}
\details{

}
\examples{
start_date <- '2018-01-01'
end_date <- '2019-12-31'

baseline  <- simulate_baseline_data(
 start_date = as.Date("2012-01-01")
 end_date = as.Date("2019-12-31")
 seasonal_pattern_n = 1
 weekly_pattern_n = 2
baseline_param_list = list(alpha = 3, beta = 0, gamma1 = 0.8, gamma2 = 0.6, gamma3 = 0.8, gamma4 = 0.4, phi = 4, shift = 29)
outbreak_param_list = list(m=2)

# )
}
